Remember to manually remove any changes to .classpath We really should put it into the .gitignore file.
I added some instanciations for the Compressor and the PDP.

added code to write the LED's on the aux drivers controls

added the shifter solenoid to the drivetrain and then made both the left
and right instances of the RSpeedController insteam of just the right.

fixed a miss-spelling of elevator in the encoder call.